Add input padding to reference convolution code
Summary:
Adding top/bottom/left/right padding to the 2d reference convolution code.

This is input padding; that is, an implied margin of 0s that actually participate in calculations, not just an offset of where the data is pulled from. If the input size is I, filter size F, padding start Ps, padding end Pe, filter stride S, the output size in a given dimension is ((I + Ps + Pe - F) / S) + 1. This actually required additional logic in the kernels for boundary conditions since it can't be expressed by changing the size/stride of tensors. Thus, input padding is more computationally complicated than output padding, which is just writing into a different offset of the output.

This is a feature that we need to support for the 3x3 kernel ImageNet 2014 which uses padding to keep the layer math consistent. Basically, the padding just adds a margin of 0s on the output.

A following diff will add this to SpatialConvolutionCaching.

Test Plan:
Padding = (0, 0, 0, 0) is tested as before, and all symmetric padding = (k, k, k, k) is tested against SpatialConvolutionMM (or rather, the code that is extracted from that module). That provides some verification that this reference implementation works, and thus, can become a reference.

A test for asymmetric padding is done by comparing the asymmetric pathway against the non-padding pathway, and by truncating the outputs or padding the inputs as needed.

Reviewed By: ntv@fb.com

Subscribers: jhj, fjargsto, tudorb, tulloch, pamelavagata, kma, ntv, soumith

FB internal diff: D1561356

Tasks: 5134296