Give up on trying to be clever with IOBuf ownership and rvalue references
Summary:We were trying to use IOBuf&& (and ThriftTensor&&, etc) in constructors to indicate
that the resulting tensor will share memory with the original buffer. This is a
convention that IOBuf itself doesn't follow; make the sharing a separate argument.
Also weaken the restriction that Storage may only share memory with a unique
IOBuf; that's not met in the real world when deserializing from Thrift (the
IOBuf in the protocol reader and the IOBuf in the Thrift structure share
memory)

Reviewed By: adamlerer

Differential Revision: D2124626

fb-gh-sync-id: dd9210b0b0bfd9cd05c74a692b2e1e135bc478e5
shipit-source-id: dd9210b0b0bfd9cd05c74a692b2e1e135bc478e5