TensorPtr
Summary:Remove thpp::Tensor aliasing. Tensors now have value semantics; making a copy
of a Tensor object will always copy the metadata (and continue to share
the data for as long as possible).

This means that a newly-created Tensor object will never refer to an existing
Lua tensor, so, for example, the following code will NOT allow you to resize
the Lua tensor:

``` lang=cpp
auto t = fblualib::luaGet<thpp::Tensor<float>>(L, -1);
t.resize({2,3});
```

but will create a new tensor t, which shares data (but not metadata) with the
Lua Torch tensor at the top of the Lua stack.

Instead, this change adds TensorPtr, which is a shared-ownership pointer to
existing tensors.

``` lang=cpp
auto t = fblualib::luaGet<thpp::Tensor<float>::Ptr>(L, -1);
t->resize({2,3});  // it's a pointer, note t->
```

The reason is that conflating aliasing and sharing metadata has confusing
semantics (what should thpp::Tensor<float> a = b; do? There are multiple
possible answers, and they all make sense in various circumstances.) It's
better to specify what you want -- if you want a pointer, ask for a pointer.

This allows us to simplify thpp::Tensor a bit -- it always points to a valid
(non-null) THFloatTensor, and that association never changes. (Unfortunately,
this means that the default constructor can't be noexcept, and so the move
constructor can't be either.)

(Eventually, we could make thpp::Tensor actually contain the THFloatTensor
object, but that requires a few changes to torch7/lib/TH)

Reviewed By: kmafb

Differential Revision: D2640819

fb-gh-sync-id: 6fcada1b55fa6bf55da5e5acbb56e0bc5b1e4218
shipit-source-id: 6fcada1b55fa6bf55da5e5acbb56e0bc5b1e4218