Support disto serialization of TH storage with non-standard allocators
Summary:
This allows disto to handle custom-allocated storage like CUDA pinned memory.
Tudor's original comment says this is unsafe because we don't know the lifetime of the allocator/context, but in all actual cases, the allocator is global, and the context has a lifetime at least until the storage is freed.

Reviewed By: soumith, colesbury

Differential Revision: D4142296

fbshipit-source-id: 87d2ab38d0be5b3d37edee56ed05127b48744b0a